Original change: https://android-review.googlesource.com/c/kernel/common/+/3216458

+++ drivers/dma-buf/heaps/system_heap.c
@@ -500,8 +500,8 @@ static const struct dma_heap_ops system_heap_ops = {
 
 static struct dma_buf *system_uncached_heap_allocate(struct dma_heap *heap,
 						     unsigned long len,
-						     unsigned long fd_flags,
-						     unsigned long heap_flags)
+						     u32 fd_flags,
+						     u64 heap_flags)
 {
 	return system_heap_do_allocate(heap, len, fd_flags, heap_flags, true);
 }
@@ -509,8 +509,8 @@ static struct dma_buf *system_uncached_heap_allocate(struct dma_heap *heap,
 /* Dummy function to be used until we can call coerce_mask_and_coherent */
 static struct dma_buf *system_uncached_heap_not_initialized(struct dma_heap *heap,
 							    unsigned long len,
-							    unsigned long fd_flags,
-							    unsigned long heap_flags)
+							    u32 fd_flags,
+							    u64 heap_flags)
 {
 	return ERR_PTR(-EBUSY);
 }
